+++ src/cjs/Circle.js
@@ -19,6 +19,7 @@
  * @modified 2026-06-10 Adding the `Circle.clone` method.
  * @modified 2026-01-13 Adding helper function `Circle.circleUtils.containsPoint` and refactored the member method `containsPoint`.
  * @modified 2026-07-03 Adding the optional `epsilon` parameter to the `Circle.containsCircle` method.
+ * @modified 2026-07-03 Fixing the `Circle.clone` method; the center had not been cloned at all, this was fixed.
  * @version  1.7.0
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
@@ -337,7 +338,7 @@ var Circle = /** @class */ (function () {
      * @memberof Circle
      **/
     Circle.prototype.clone = function () {
-        return new Circle(this.center, this.radius);
+        return new Circle(this.center.clone(), this.radius);
     };
     Circle.fromICircle = function (obj) {
         return new Circle(new Vertex_1.Vertex(obj.center), obj.radius);