+++ src/ts/Circle.ts
@@ -18,6 +18,7 @@
  * @modified 2026-06-10 Adding the `Circle.clone` method.
  * @modified 2026-01-13 Adding helper function `Circle.circleUtils.containsPoint` and refactored the member method `containsPoint`.
  * @modified 2026-07-03 Adding the optional `epsilon` parameter to the `Circle.containsCircle` method.
+ * @modified 2026-07-03 Fixing the `Circle.clone` method; the center had not been cloned at all, this was fixed.
  * @version  1.7.0
  **/
 
@@ -395,7 +396,7 @@ export class Circle implements IBounded, ICircle, Intersectable, SVGSerializable
    * @memberof Circle
    **/
   clone(): Circle {
-    return new Circle(this.center, this.radius);
+    return new Circle(this.center.clone(), this.radius);
   }
 
   static fromICircle(obj: ICircle): Circle {